polished detail pane
Polished up the detail pane so that the grid layout stretches to fit
rather then looking sad and off to the right. Added buttons for tracking
(notifications will come later). The map images all load (and are
statically sized, dynamic scaling is next). Clicking on the image opens
map activities and they are all at the correct locations.

This version is stable.