Notifications set correctly.
Added real notifications. They are based on the time loaded from the
server. They do not update. They are somewhat accurate now, but
background services that keep checking the times need to be added later
on.
Added support for the buttons and error handling so that you cannot
track a zero or errored shuttle