FusedLocationAPI, load time storage, bugfixes.
Now uses the googleApiClient and FusedLocationAPI to determine what
location provider is correct for the current setting. GPS for high
accuracy outdoors or when moving, network location for indoors.
Also stores the last time the times were loaded/location updated so it
does not update every time the fragment is started.